cast pid to long long as %jd isn't supported by windows

+++ tools/joypad_mapping/joypad_mapping.c
@@ -542,7 +542,7 @@ int main(int argc, char **argv)
 	/* mandatory for attribute cleanup not to segfault in error cases */
 	reset_joystick_config(&joystick_config);
 	progname = basename(argv[0]);
-	printf("%s[%jd] starting\n", progname, (intmax_t)getpid());
+	printf("%s[%lld] starting\n", progname, (long long)getpid());
 	if (argc == 3) {
 		if (strcmp(argv[1], "--test") != 0)
 			usage(progname);
